+++ Events/LocationEvent.cs
@@ -34,6 +34,7 @@ namespace EddiEvents
             VARIABLES.Add("longitude", "The longitude of the commander (if on the ground)");
             VARIABLES.Add("latitude", "The latitude of the commander (if on the ground)");
             VARIABLES.Add("population", "The population of the system to which the commander has jumped");
+            VARIABLES.Add("factions", "The factions in the system (this is a list of faction objects)");
         }
 
         public string system { get; private set; }
@@ -72,6 +73,8 @@ namespace EddiEvents
 
         public decimal? latitude { get; private set; }
 
+        public List<Faction> factions { get; private set; }
+
         // These properties are not intended to be user facing
         public long? systemAddress { get; private set; }
         public long? marketId { get; private set; }
@@ -83,7 +86,7 @@ namespace EddiEvents
         public StationModel stationModel { get; private set; } = StationModel.None;
         public BodyType bodyType { get; private set; } = BodyType.None;
 
-        public LocationEvent(DateTime timestamp, string system, decimal x, decimal y, decimal z, long systemAddress, string body, BodyType bodytype, bool docked, string station, StationModel stationtype, long? marketId, Superpower allegiance, string faction, Economy economy, Economy economy2, Government government, SecurityLevel security, long? population, decimal? longitude, decimal? latitude) : base(timestamp, NAME)
+        public LocationEvent(DateTime timestamp, string system, decimal x, decimal y, decimal z, long systemAddress, string body, BodyType bodytype, bool docked, string station, StationModel stationtype, long? marketId, Superpower allegiance, string faction, Economy economy, Economy economy2, Government government, SecurityLevel security, long? population, decimal? longitude, decimal? latitude, List<Faction> factions) : base(timestamp, NAME)
         {
             this.system = system;
             this.x = x;
@@ -105,6 +108,7 @@ namespace EddiEvents
             this.population = population;
             this.longitude = longitude;
             this.latitude = latitude;
+            this.factions = factions;
         }
     }
 }